[SASS-2955] Demo
If I understand, the main intent is to reduce the running time of the integration tests.

Typically, we have four 'user scenarios' for each: agent or individual, English or Welsh.

With these changes, we reduce the slow integration test to one scenario, and move the four scenarios to unit tests, which assert the generated form in a new way.

My main concern is that the integration test only assesses the status code, and not the form itself; given that we've already paid for the page, it would cost us nothing to assert the form as we used to.

If we did this, then the *Spec would be a realistic test (always with a single user scenario, of course) whereas the companion *ViewSpec would exercise the four (and more, if needed) variations.

And we could share as much as possible - preferably the user scenarios themselves, even keeping the user scenario sequence in the *Spec, but with only one member.

Granted, I didn't attempt this much; but I did show how the form assertions could be restored.