[SASS-2955] Proposal: More transparent tests.
This is quite a deviation, and I'm not sure if it's a clear improvement.

While it does have duplication, it's arguably more transparent.

It seems that we have 4 'user scenarios' at most: individual / agent and English / Welsh.

It should be clear when each is present.

Given the prevalence of the user scenario approach, I wouldn't necessarily recommend this change for this ticket, or any other: I just wanted to get your feedback on it.