SASS-7127: Fix Journey1 CYA (#364)
Whenever we answered No we never send any numeric values to IFS,

therefore we were not able to figure out on load prior data if user has already submitted value. One way to solve that, and I think that's what IFS wants is that if we answered No, we will send 0.0 values there.

It also solved the problem that if the value was e.g. 5.0 before, if I answer No, I will override my obsolete values.